requirements.txt
anyio==4.4.0
boto3==1.34.153
debugpy>=1.8.0,<2.0.0
fastapi>=0.112.0,<1.0
httpx==0.27.0
hypothesis==6.108.5
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
langchain==0.2.12
langchain_anthropic==0.1.22
langchain_community==0.2.11
langchain_core==0.2.28
langchainhub>=0.1.20,<1.0
langchain_openai==0.1.20
langchain-pinecone>=0.1.3,<2.0
langsmith==0.1.96
mmh3==4.1.0
numpy>=1.26.0,<2.0.0
openai==1.44.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
protobuf<5.0,>=3.19
psutil==6.0.0
pydantic==2.8.2
PyPika==0.48.9
pytest==8.3.2
python-dotenv==1.0.1
bcrypt==4.0.1
PyYAML==6.0.1
PyYAML==6.0.1
rich==13.7.1
starlette<0.38.0,>=0.37.2
tenacity==8.5.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.5
pinecone-client>=5.0.1,<6.0
python-dotenv==1.0.1
redis>=5.0.8,<6.0
langchain-postgres>=0.0.9,<1.0
SQLAlchemy>=2.0.31,<3.0
python-jose>=3.3.0,<4.0
psycopg2-binary>=2.9.9,<3.0
psycopg[binary]>=3.2.1,<4.0
passlib>=1.7.4,<2.0
python-multipart>=0.0.9,<1.0
requests>=2.32.3,<3.0
alembic>=1.13.2,<2.0
replicate>=0.31.0,<1.0
slowapi>=0.1.9,<0.2.0
uuid>=1.30.0,<2.0
sendgrid>=6.11.0,<7.0
toml>=0.10.2,<1.0
sentry-sdk>=2.13.0,<3.0
loguru>=0.7.2,<1.0
termcolor>=2.4.0,<3.0
docstring_parser>=0.16,<1.0
pypdf>=4.3.1,<5.0
pillow>=10.4.0,<11.0
google-cloud-storage>=2.18.2,<3.0
pygame>=2.6.0,<3.0
pytest-asyncio>=0.24.0,<0.25.0
tiktoken>=0.5.0,<1.0